feat: just dump everything lol
Not my usual standard for commits, but I've been experimenting with
different things on my laptop and I want to access this on my desktop as
well.

I'm not sure about my directory structure yet.

I ended up solving puzzle 1 more or less by accident using Nushell. I
was experimenting with the input data to see if there were any edge
cases I didn't consider, and then I though "let's see how hard it would
be to do in Nu". Turns out, once again, that nu is amazing.

The Gleam solution for the first puzzle is close, but the number isn't
quite right. I do have tests, which is nice. And it streams the file one
line at a time using a recursive function, which is pretty sick.